Moving WhatsNew component to its own package.
