+++ packages/components/src/index.js
@@ -8,4 +8,3 @@ export { default as Ribbon } from './ribbon';
 export { default as RootChild } from './root-child';
 export { default as ScreenReaderText } from './screen-reader-text';
 export { default as Suggestions } from './suggestions';
-export { default as WhatsNewGuide } from './whats-new';
