+++ packages/whats-new/package.json
@@ -1,0 +1,51 @@
+{
+	"name": "@automattic/whats-new",
+	"version": "1.0.0",
+	"description": "A WordPress Guide component that highlights recently-added WordPress.com features.",
+	"main": "dist/cjs/index.js",
+	"module": "dist/esm/index.js",
+	"sideEffects": false,
+	"keywords": [
+		"wordpress"
+	],
+	"author": "Automattic Inc.",
+	"contributors": [],
+	"homepage": "https://github.com/Automattic/wp-calypso",
+	"license": "GPL-2.0-or-later",
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/Automattic/wp-calypso.git",
+		"directory": "packages/your-package"
+	},
+	"publishConfig": {
+		"access": "public"
+	},
+	"bugs": {
+		"url": "https://github.com/Automattic/wp-calypso/issues"
+	},
+	"files": [
+		"dist",
+		"src"
+	],
+	"scripts": {
+		"clean": "npx rimraf dist",
+		"prepublish": "yarn run clean",
+		"prepare": "transpile"
+	},
+	"dependencies": {
+		"@automattic/calypso-analytics": "^1.0.0-alpha.1",
+		"@babel/runtime": "^7.12.5",
+		"@automattic/react-i18n": "^1.0.0-alpha.0",
+		"@wordpress/components": "^12.0.7",
+		"wpcom": "^5.4.2"
+	},
+	"peerDependencies": {
+		"react": "^16.8",
+		"react-dom": "^16.8"
+	},
+	"devDependencies": {
+		"wpcom-proxy-request": "^6.0.0",
+		"react": "^16.12.0",
+		"react-dom": "^16.12.0"
+	}
+}
